spec fixes, added execution context for iomanagers, abstracted iomanagers, added license

+++ src/test/scala-2.11/spec/DBSpec.scala
@@ -1,22 +1,27 @@
 package spec
 
-import cortex.io.Cortex
-import cortex.util.test
+import java.io.File
+
+import cortex.db.SqlDB
 import org.scalatest.matchers.{MatchResult, Matcher}
-import org.scalatest.{FlatSpec, Matchers}
+
 import scalikejdbc._
 
 /**
- */
+  */
 class DBSpec extends BaseSpec {
-  import cortex.db.DB._
+
+  import cortex.db.SqlDB._
+
   case class Person(id: Long,
                     firstName: String,
                     lastName: String,
                     email: String,
                     hash: String)
+
   object Person extends SQLSyntaxSupport[Person] {
     override val tableName = "persons"
+
     def apply(rs: WrappedResultSet): Person = new Person(
       rs.long("id"),
       rs.string("first_name"),
@@ -26,48 +31,45 @@ class DBSpec extends BaseSpec {
     )
   }
 
-  @test object app extends Cortex {
-    override def port = 9997
+  override def configFile = Option(new File("src/test/config.txt"))
 
-    override def controllers = Seq()
-    override def views = Seq()
-    override def dbConnection = DBConnection(
-      "localhost:3306/conf",
-      config.user,
-      config.password
-    )
-  }
-
-  app.hashCode()
-
-  it should "validate test information in the database" in {
-    sql"""
-      INSERT INTO persons(first_name, last_name, email, hash)
-      VALUES("test", "test", "test@test.com", "test")
-    """.update().apply()
-
-    val person: Option[Person] = sql"""
-      SELECT * FROM persons
-      WHERE email='test@test.com'
-    """.map(rs => Person(rs)).single().apply()
+  override def dbConnection = DBConnection(
+    "localhost:3306/conf",
+    config.user,
+    config.pass
+  )
 
 
-    sql"""
-      DELETE FROM persons
-      WHERE email='test@test.com'
-    """.update().apply()
+  if (SqlDB.isInitialized) {
+    it should "validate test information in the database" in {
+      sql"""
+        INSERT INTO persons(first_name, last_name, email, hash)
+        VALUES("test", "test", "test@test.com", "test")
+      """.update().apply()
 
-    val bePerson =
-      new Matcher[Person] {
-        def apply(left: Person) = {
-          MatchResult(
-            left.email.equals("test@test.com"),
-            s"$left did not match test case",
-            s"$left matched test case"
-          )
+      val person: Option[Person] =
+        sql"""
+          SELECT * FROM persons
+          WHERE email='test@test.com'
+        """.map(rs => Person(rs)).single().apply()
+
+      sql"""
+        DELETE FROM persons
+        WHERE email='test@test.com'
+      """.update().apply()
+
+      val bePerson =
+        new Matcher[Person] {
+          def apply(left: Person) = {
+            MatchResult(
+              left.email.equals("test@test.com"),
+              s"$left did not match test case",
+              s"$left matched test case"
+            )
+          }
         }
-      }
 
-    person.get should bePerson
+      person.get should bePerson
+    }
   }
 }
